cloud/weather-dashboard/requirements.txt
flask
requests
python-dotenv
google-cloud-storage
